refactor: improve fileutil test organization by splitting into focused test files
Refactored the large fileutil_test.go (1603 lines) into smaller, focused test files:
- core_utils_test.go: Basic utility tests
- config_test.go: Configuration-related tests
- git_operations_test.go: Git-related functionality tests
- filtering_test.go: File filtering tests
- file_processing_test.go: File processing and error handling tests
- project_context_test.go: Context gathering tests

This improves maintainability by organizing tests by functionality.
