+++ .gitignore
@@ -43,8 +43,6 @@ PLAN.md
 test-results/
 thinktank-output/
 architect_output/
-TODO.md
-PLAN.md
 *-TASK.md
 *-PLAN.md
 CONSULT-REQUEST.md
