+++ internal/fileutil/core_utils_test.go
@@ -1,0 +1,129 @@
+// internal/fileutil/core_utils_test.go
+package fileutil
+
+import (
+	"strings"
+	"testing"
+)
+
+func TestEstimateTokenCount(t *testing.T) {
+	tests := []struct {
+		name     string
+		input    string
+		expected int
+	}{
+		{
+			name:     "Empty string",
+			input:    "",
+			expected: 0,
+		},
+		{
+			name:     "Single word",
+			input:    "hello",
+			expected: 1,
+		},
+		{
+			name:     "Two words",
+			input:    "hello world",
+			expected: 2,
+		},
+		{
+			name:     "Multiple spaces",
+			input:    "hello  world",
+			expected: 2,
+		},
+		{
+			name:     "With newlines",
+			input:    "hello\nworld",
+			expected: 2,
+		},
+		{
+			name:     "With tabs",
+			input:    "hello\tworld",
+			expected: 2,
+		},
+		{
+			name:     "Mixed whitespace",
+			input:    "hello\n \t world",
+			expected: 2,
+		},
+	}
+
+	for _, test := range tests {
+		t.Run(test.name, func(t *testing.T) {
+			result := estimateTokenCount(test.input)
+			if result != test.expected {
+				t.Errorf("estimateTokenCount(%q) = %d, expected %d", test.input, result, test.expected)
+			}
+		})
+	}
+}
+
+func TestCalculateStatistics(t *testing.T) {
+	input := "Hello\nWorld\nThis is a test."
+	expectedChars := len(input)
+	expectedLines := 3
+	expectedTokens := 6 // "Hello", "World", "This", "is", "a", "test."
+
+	chars, lines, tokens := CalculateStatistics(input)
+
+	if chars != expectedChars {
+		t.Errorf("Character count: got %d, want %d", chars, expectedChars)
+	}
+
+	if lines != expectedLines {
+		t.Errorf("Line count: got %d, want %d", lines, expectedLines)
+	}
+
+	if tokens != expectedTokens {
+		t.Errorf("Token count: got %d, want %d", tokens, expectedTokens)
+	}
+}
+
+func TestIsBinaryFile(t *testing.T) {
+	tests := []struct {
+		name     string
+		content  []byte
+		expected bool
+	}{
+		{
+			name:     "Empty content",
+			content:  []byte{},
+			expected: false,
+		},
+		{
+			name:     "Text content",
+			content:  []byte("This is a text file with some content.\nIt has multiple lines."),
+			expected: false,
+		},
+		{
+			name:     "Content with null byte",
+			content:  []byte("This has a null byte.\x00Right there."),
+			expected: true,
+		},
+		{
+			name:     "Content with many non-printable characters",
+			content:  []byte{0x01, 0x02, 0x03, 'H', 'e', 'l', 'l', 'o', 0x04, 0x05},
+			expected: true,
+		},
+		{
+			name:     "Content with few non-printable characters",
+			content:  []byte("Hello\nWorld\tThis is a test with a bell sound: \a"),
+			expected: false,
+		},
+		{
+			name:     "Very long text content",
+			content:  []byte(strings.Repeat("This is a test. ", 100)),
+			expected: false,
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			result := isBinaryFile(tt.content)
+			if result != tt.expected {
+				t.Errorf("isBinaryFile() = %v, want %v", result, tt.expected)
+			}
+		})
+	}
+}
